+++ app/Console/Commands/AdsFetchHosts.php
@@ -27,9 +27,11 @@ use Adshares\Ads\Driver\CommandError;
 use Adshares\Ads\Entity\Broadcast;
 use Adshares\Ads\Exception\CommandException;
 use Adshares\Adserver\Console\Locker;
+use Adshares\Adserver\Http\Response\InfoResponse;
 use Adshares\Adserver\Models\NetworkHost;
 use Adshares\Common\Exception\RuntimeException;
 use Adshares\Network\BroadcastableUrl;
+use Adshares\Supply\Application\Dto\Info;
 use Adshares\Supply\Application\Service\DemandClient;
 use Adshares\Supply\Application\Service\Exception\UnexpectedClientResponseException;
 use Illuminate\Support\Facades\Log;
@@ -42,18 +44,14 @@ class AdsFetchHosts extends BaseCommand
     const BLOCK_TIME = 512;
 
     /**
-     * Period is seconds which will be searched for broadcast
+     * Period in seconds which will be searched for broadcast
      */
     const PERIOD = 43200;//12 hours = 12 * 60 * 60 s
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $signature = 'ads:fetch-hosts';
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $description = 'Fetches Demand AdServers';
 
     /** @var DemandClient */
@@ -66,9 +64,6 @@ class AdsFetchHosts extends BaseCommand
         parent::__construct($locker);
     }
 
-    /**
-     * @param AdsClient $adsClient
-     */
     public function handle(AdsClient $adsClient): void
     {
         if (!$this->lock()) {
@@ -96,11 +91,6 @@ class AdsFetchHosts extends BaseCommand
         $this->info('Finished command '.$this->signature);
     }
 
-    /**
-     * @param int $timeNow
-     *
-     * @return int
-     */
     private function getTimeOfFirstBlock(int $timeNow): int
     {
         $timeStart = $timeNow - self::PERIOD;
@@ -115,10 +105,6 @@ class AdsFetchHosts extends BaseCommand
         return $timeBlock;
     }
 
-    /**
-     * @param AdsClient $adsClient
-     * @param string $blockId
-     */
     private function handleBlock(AdsClient $adsClient, string $blockId): void
     {
         try {
@@ -138,9 +124,6 @@ class AdsFetchHosts extends BaseCommand
         }
     }
 
-    /**
-     * @param $broadcast
-     */
     private function handleBroadcast(Broadcast $broadcast): void
     {
         $address = $broadcast->getAddress();
@@ -152,6 +135,8 @@ class AdsFetchHosts extends BaseCommand
 
             $info = $this->client->fetchInfo($url);
 
+            $this->validateInfo($info, $address);
+
             Log::debug("Got {$url->toString()}");
 
             $host = NetworkHost::registerHost($address, $info, $time);
@@ -160,6 +145,25 @@ class AdsFetchHosts extends BaseCommand
         } catch (RuntimeException|UnexpectedClientResponseException $exception) {
             $url = $url ?? '';
             Log::debug("[$url] {$exception->getMessage()}");
+        }
+    }
+
+    private function validateInfo(Info $info, string $address): void
+    {
+        if (InfoResponse::ADSHARES_MODULE_NAME !== $info->getModule()) {
+            throw new RuntimeException(sprintf('Info for invalid module: %s', $info->getModule()));
+        }
+
+        $adsAddress = $info->getAdsAddress();
+
+        if (!$adsAddress) {
+            throw new RuntimeException('Info has empty address');
+        }
+
+        if ($adsAddress !== $address) {
+            throw new RuntimeException(
+                sprintf('Info has different address than broadcast: %s !== %s', $adsAddress, $address)
+            );
         }
     }
 }