+++ database/factories/NetworkHost.php
@@ -23,6 +23,7 @@ use Adshares\Common\Domain\ValueObject\AccountId;
 use Adshares\Common\Domain\ValueObject\Email;
 use Adshares\Common\Domain\ValueObject\Url;
 use Adshares\Supply\Application\Dto\Info;
+use Adshares\Supply\Application\Dto\InfoStatistics;
 use Faker\Generator as Faker;
 
 $factory->define(
@@ -41,7 +42,7 @@ $factory->define(
 
         $host = 'https://'.parse_url($faker->url, PHP_URL_HOST);
         $info = new Info(
-            'ADSERVER',
+            'adserver',
             $faker->domainWord,
             '0.1',
             new Url($host),
@@ -49,10 +50,15 @@ $factory->define(
             new Url($host.'/policies/privacy.html'),
             new Url($host.'/policies/terms.html'),
             new Url($host.'/adshares/inventory/list'),
+            new AccountId('0001-00000004-DBEB'),
             new Email($faker->companyEmail),
             'PUB',
             'ADV'
         );
+
+        $info->setDemandFee(0.01);
+        $info->setSupplyFee(0.01);
+        $info->setStatistics(new InfoStatistics(1, 1, 1));
 
         return [
             'address' => $faker->randomElement($addresses),